auth: google auth is separated for web-app and android
There will be multiple GoogleAuthProvider instances, each for a
different purpose. That is because OAuth 2.0 works in a different
way for web applications and for smart phone applications.

Thus, the instance googleAuthProvider is renamed to googleWebAuth
and another similar instance named googleAndroidAuth is added.
The instances IDs differ for unique referencing from clients:
google-webapp and google-android.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>