FCMSender: new class
FCMSender class provides a process of dispatching FCMMessage instances
via a given FCMClient. It implements defering of messaging for cases
when the FCMClient fails to deliver a message.

It handles also partial undelivery (to some recipients), failing delivery
(recipient does not exist), renewing of a FCMTokenID. Undelivered messages
are resended by using the ExponentialBackOff delay algorithm.

Signed-off-by: Peter Kubov <xkubov06@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>