AuthRestHandler: publish OAuth2 redirect URI to client
The client needs the right redirect URI for successful login via OAuth 2.0.
The server can provided in the auth methods listing. Propagate the redirect
URI from the OAuth2AuthProvider.

This also allows the client to better filter the OAuth 2.0 methods it can
use because for different client applications there can be different
requirements of the target redirect URI.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>