xmlui: Serializing: provide gateway timezone to clients
Until now, the serializing of Gateway instances was serving a constant "0"
time zone specification. Now, we can grap the offset (in seconds) from the
timeZone property of the gateway instance.

Also, add short name of the time zone as timezone_name to the output. We
should not put there display name as XML-UI is not aware of the area
specification. The short name is usually just in form GMT+01:00, etc.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>