DeviceRestHandler: asynchronous operations are redirected via 202
The responses to devices.discover and devices.unpair are asynchronous and return
redirect to a specific location where the particular work instance is accessible.
This is however bad for the browser that does not interpret the Location header
on POST and DELETE requests.

Change the return type to 202 Accepted which says that the operation is in progress
and return a JSON body containing the location field. This way, the client can
access the operation progress on its own easily.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>