tester: new module sensors
The module provides an access to sensors history. It is possible to
configure the sensor history requests via time range, aggregation
and aggregation interval.

The sensor data history is displayed in a simple chart generated
via the chartjs [1] library. The implementation is however not very
clever and should be improved later to just redraw the existing
canvases instead of rebuilding them.

[1] http://www.chartjs.org/

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>