GatewayWatcher: new class
GatewayWatcher reacts on events coming from the GatewayCommunicator. It sends
FCM notifications for all the three event types.

There is one implementation catch to be considered. The FCMSender manages
database transactions internally while handling FCMResponse. Every single
FCMResponse should be done in a separate transaction (at least). However,
the GatewayWatcher also manages a transaction. As we do not support inner
transactions, we have to first do the GatewayWatcher transaction and then
call the FCMSender::send().

Another catch, the FCMSender blocks us while processing the gateway events.
For each message, it does a round trip to the Firebase Cloud Messaging service
which might be slow. It is desirable to move all FCMSender processing logic
to a separate thread in the future.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>