ThreadLocalTransactionManager: avoid double-delete call
Calling start() on the ThreadLocalTransactionManager should lead either to
a new transaction or to an error. Otherwise, it is possible to create a
nested transaction accidently that is actually the same as the outer one.
Calling delete inside the inner transaction might lead to a segfault in
the outer transaction when using the BEEEON_TRANSACTION macros.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>