tester: implement the firebase-messaging-sw.js via CPSP
The class FcmswHandler implemented via the CPSP system is a standalone generated
web page serving the background handler for the Firebase Cloud Messaging service.
Currently, it is mapped as /tester/firebase-messaging-sw.js.

The FcmswHandler inherits from FCMReceiverConfigurable because there is currently
no easy way how to inject setters into the CPSP page (at least it seems to...).
The FcmswHandler needs to know the sender ID to make the background handler to
work properly.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>