GatewayConnection: fix handling of opcode CLOSE
The current implementation handles WebSocket opcodes in a wrong way.
An opcode is not a bitmap, it is a value. By testing via the logical
AND operator (&), we close on opcodes:

* CLOSE = 0x08
* PING  = 0x09
* PONG  = 0x0a

because all of them satisfies the mask. We however want to close on
opcode CLOSE only. Also, the flags field contains also other flags
then just the opcode and thus we have to mask unrelated bits out.

Note that we are not handling other opcodes well and thus not conforming
to the WebSocket standard right now.

Fixes: 220e617aeeec ("GatewayConnection: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>