PocoAbstractDao: log statement exceptions with bound params
If a statement fails, the following change would allow to introspect
what were the actual parameters bound to the statement. The class
IntrudingBinder is used to rebound all the IN parameters of the
failed statements and this way to extract their values.

Also, the exceptions are logged via a single logger call instead
of two before where the query and the exception have been logged
independently.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>