FacebookAuthProvider: implement verifyAuthCode by access-token
The Facebook login is already divided into three steps. Conversion of the
authorization code into an access token, verification of the access token
and fetching of user data. The first step protects the user from stealing
the access token while sending it to the server.

The verifyAuthCode(AccessTokenCredentials) implementation avoids the first
step. This makes the login slightly more vulnerable to some attacks. However,
the login should be secure enough from the point of view of the server.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>