AuthServiceImpl: do not use NotificationDispatcher
We have to remove NotificationDispatcher and related stuff from the
AuthServiceImplTest to build and pass. The test was also testing that
the notification is fired by setting the m_lastIdentity. This behaviour
is not preserved and we are not going to fix this.

If needed, the same notification principle can be implemented via the
IdentityListener.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>