xmlui: Serializing: fix device type info
The attribute displayName is wrong. Using of methods name, vendor and
displayName and displayVendor is currently floating in the air. The goal
is to provide device type info to the client. Currently, the values coming
from methods name() and vendor() are empty and never filled. Thus, prefer
methods displayName() and displayVendor().

Also, change the attributes' names. The attribute @displayName is wrong.
It is not a displayName but full name of the type. This can however be
constructed from type name and vendor.

The attribute @name is also wrong, it is now renamed to @type_name which
discribes the purpose better. Currently, it is unused as there is a duplicate
attribute @name defined earlier.

Fixes: d662a41b1f0b ("xmlui: Serializing: changes in DeviceInfo")
Fixes: 2eae6567e0c5 ("DeviceInfoProvider: implement findByNameAndVendor using match()")

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>