DeviceInfo: get rid of display* members
The displayName and displayVendor are now replaced with name and vendor.
The transition is done. No more display* things are needed here as the
devices.xml gives only name and vendor. Other information are separated
for matching strategy. The frontend can server display* values but via
the l10n Translator. The translation would be based on the name and vendor
values.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>