PocoRestRequestFactory: create handler in tmp variable
The handler is first instantiated and then returned. All this must
be wrapped in the try ... catch to avoid leaks in case of something
really terrible (and here we are on the shooting range).

This allows to move some arguments passing via constructor into
setters and thus to simplify the code.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>