FailingTCPServerConnection: new class
In case, when a socket connection fails early (client closes connection)
inside the XmlRequestHandlerFactory::createConnection(), the factory
returns NULL. This seems however to be insane and the TCPServer may
behave wrong. Thus, instead return a new instance of the introduced
class FailingTCPServerConnection that properly starts but immediately
throws ConnectionAbortedException.

The issue could probably lead to hang of the server. It was discovered
during integration tests - by repeatedly starting all tests over and
over as

 $ while true; do prove t/xmlui/t*.py; done

It always hangs at some place without a reason with NullPointerException.
and a message about "connection reset by peer". With this patch, this
issue does not happen.

Fixes: 2621f46a74ed ("XmlRequestHandler: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>