PocoSQLSensorHistoryDao: pass timestamp in microseconds
All sensoric data are now being stored with microseconds resolution.
This should avoid several constraint violation errors like:

 [Error] Database Exception: ODBC State mentException
 [0] SQL state: 23505, Native error: 7
 ERROR: duplicate key value violates unique constraint "sensor_history_recent_pk";
 Error while executing the query
 SELECT beeeon.sensor_history_recent_insert($1::bigint, beeeon.to_device_id($2::
 		numeric(20, 0)), $3::smallint, beeeon.as_utc_timestamp($4), $5)

It was happening due to equal or too-close-to-each-other timestamps.
However, this is not a solution of the problem "too many data coming
shortly after each other"! This patch allows to fill the database very
quickly if somebody sends that many data. But anybody could try to send
a lot of data with timestamps which are very far from each other and we
would be not able to catch it anyway.

This mechanism seems to be better right now than loosing data because of
just wrong testing gateway settings. Some explicit limiting mechanism is
needed to filter those attempts on the application layer.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>