DeviceServiceImpl: add missing policy call
Because we are fetching devices for a single gateway here, we can perform
a check for access to the gateway itself first. This can prevent as for
losing all the devices from database in case of an invalid access attempt.

Note that this does not fix any security issue as the access is checked
later on per device basis. Here, we just safe some calls in case of
insufficient permissions.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>