PocoRestRequestHandler: simplify creation of translator
Determining of locale does not have to be done explicitly by testing
session existence in the PocoRestRequestHandler. This logic is already
contained inside RestFlow::locale() method. Thus, use this method and
simplify the setupLanguage() routine. This leads to change in the
logging message as we does not clearly know whether the "user" locale
or "http" locale was selected. But this is not so important information.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>