DeviceStatus: rework activeSince property
The activeSince does not exist anymore at the model level. Instead,
each device has its state and the timestamp of last change. The
combination covers the activeSince logic and extending it a bit.
It would be now possible to determine when a device has been
deactivated. Also, the pending states can now be used.

Both the REST-UI and XML-UI now reports the more detailed state
of the device together with the last-changed timestamp.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>