DeviceUnpairWork: remove and reimplement in service
The unpair process will be done differently. First, we have
to get rid of the overcomplex DeviceUnpairWork implementation.
The GWS callback is implemented as lambda directly inside the
DeviceServiceImpl. The unregister operation does not return
Work instance anymore. The progress is reported via the Device
entity itself and its DeviceStatus property.

The lambda functions are mutable to simplify passing context
(DeviceDao::Ptr). This should not be misused for changing the
state of the lambda otherwise race conditions are possible.

The logging of the unpairing process is improved to match the
same principles used in the pairing process.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>