DeviceMatchingTester: new class
DeviceMatchingTester is a Tool that works interactivelly by using
a Console. It is designed to work from the command line and thus
the only recommended Console implementation for now is StdConsole.

The purpose is testing of matching of devices against devices.xml. The
advantage of this approach is that it can be started with the exactly
same configuration as the server is started.

During development, run as:

 $ beeeon-server -c conf/testing-startup.ini \
	 -Dapplication.di.runner=deviceMatchingTester \
	 -Dspec.devices.xml=`pwd`/conf/devices.xml

or in production environment as:

 $ beeeon-server -c /etc/beeeon/server/server-startup.ini \
	 -Dapplication.di.runner=deviceMatchingTester

Possible queries:

 > list
 > match Thermona "Zone 1"
 > {"vendor": "Thermona", "product_name": "Zone 1"}

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>