GWSDeviceServiceImpl: check device type matches
Previously, the missing device type was detected at Dao level. This is
wrong. TheDao level performs more a kind of an assert and not test of
valid inputs. Change the GWSDeviceServiceImpl to throw NotFoundException
for this case. It means that if the new reported device does not match
any known device type (from devices.xml), the NotFoundException is thrown.
The non-matching specification is also reported.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>