TimeZoneProvider: provide method findByOffset()
We may need to lookup time zones by the UTC offset. In such case,
the expected results are usually zones names like GMT+01:00, GMT-05:00,
etc. This was not easily possible with the previous API that could only
provide a specific time zone by ID or a list of time zones which
are suitable for selection by user (contains location information).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>