FCMTokenServiceImpl: fix handling of existing token
When registering an already existing token, the register service call
fails and the server returns HTTP status 500. This is not what we
would like to see.

This changes the behaviour to be silent while registering a duplicate
token. However, if such token already exists but for a different user
(this should never happen however), we fail with InvalidArgumentException
(thus 400 Bad request).

Fixes: cecc4c3b0f7a ("FCMTokenServiceImpl: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>