AsyncGatewayRPC: always call resultCall from a different thread
There is a case when the resultCall is always invoked from the same thread
it is created in. This means that when doing e.g.:

 Event event;

 m_asyncRPC->sendListen([...]() {
 	event.set();
 }, ...);

 event.wait();

would lead to deadlock. A similar problem arises with transactions.

Such sitation occures when the target gateway is not registered with
the GW Server or in other exceptional situations. We use the embedded
timer to schedule calling of the resultCall to be invoked in another
thread as soon as possible.

Fixes: f6beaaf99168 ("AsyncGatewayRPC: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>