OAuth2AuthProvider: refactoring of HTTP methods
The convertResponseToString() method is removed because the HTTPEntireResponse
class can convert the response to a string. Also the connectSecure() method
is deleted because the connection is created in method HTTPUtil::makeRequest().
It is not necessary to set URI to request because the HTTPUtil's methods take care
about it. Simplify the makeRequest() method that only uses HTTP::makeRequest()
to send the request.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>