Control: redesign and simplify
The original ideas behind Control and its stability were too
complicated for implementation and actually over-designed its
purpose. Here, we transform the Control class to a simplier
approach. The Control is still an AbstractModule but now it
has 2 properties:

* requested value

  The requested value represents a change request. If some originator
  (usually a user) asks to change state of the associated device module,
  the requested value is filled with the _target value_ and _requested at_
  timestamp.

  When the request is accepted by the appropriate gateway, the state of
  the requested value is updated - the field _accepted at_ contains the
  timestamp of the acceptation.

  Finally, a success or a failure is represented by _finished at_
  timestamp and the _result_ property.

* recent value

  The recent value is represents the last value reported by the target
  device module. It is the same value as reported from simple sensors.

Database representation is now made of two tables: controls_fsm and
sensor_history_recent. The sensor_history_recent contains the history
of values reported by the actor while the controls_fsm the history of
requests and their results. The most recent value in controls_fsm and
sensor_history_recent does not have to match. However, it is expected
that after a successful request, the actor at least once reports its
value.

The Control entity is stored in the controls_fsm table. There is always
exactly zero or one active request (row with finished_at IS NULL) in the
table for a single control. If there is no such role, we represent such
Control with requestedAt() (and originator()) property set to NULL.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>