TypesSAXHandler: allow specify multiple units
This is a preparation step to support scaling of units for certain
types. A type can sometimes be represented by different units. The
pressure can be expressed in Pa, hPa, kPa, however it can also be
expressed in bar, mbar, ... Allow define such units for each type
with the conversion formulas.

The TypesSAXHandler currently handles:

* <unit> element (backwards compatible, stable)
* first <unit> element inside <units> element

So, currently, we are not able to parse the alternative units, we
are just not failing on parsing those.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>