WebSocketRequestHandlerFactory: verify gateway peers
The method createRequestHandler now creates a GatewayPeerVerifier instance
for each request. Because the Poco::Net::HTTPServerRequest does not provide
any means to access the underlying socket, we have to break the polymorphism
here and force to access the HTTPServerRequestImpl. This operation is however
guaranteed as there is currently no other HTTPServerRequest available.

Later, when the register message is received by the WebSocketRequestHandler,
it calls the verifyPeer on the current GatewayPeerVerifier instance to test
whether the new incoming gateway is the gateway we are expecting it to be.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>