install: avoid duplicating release info in rpm
On fighting with cpack to build RPM as expected...
The previous fix to include the release part of the version
also in the file name led to the following situation:

$ yum info beeeon-server
...
Version     : 2018.1.1.0.rc1
Release     : 0.rc1
...

The "0.rc1" entry is duplicated here. It is more clear when
using the rpm:

$ rpm -iq beeeon-server
beeeon-server-2018.1.1.0.rc1-0.rc1.x86_64
                       ^^^^^ ^^^^^

Thus, we change again the way of generating the RPM file name.
Hopefully, it is now correct.

Fixes: 4d4546cc2dbe ("install: change name of RPM package")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>