PocoAbstractDao: inject profiler into queries
All dao implementations can now be profiled by a QueryProfiler.
The PocoAbstractDao injects the QueryProfiler into each registered
query.

Note that there may be created a dependency on QueryProfiler
that is automatically unresolvable by the DI. Thus, it is necessary
to provide the cleanup hook to clear the queries (and thus the
profiler references).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>