SocketGatewayPeerVerifierFactory: fix method extractPeerCert
The original method extractPeerCert() assumed calling asSecure() to convert
a StreamSocket instance to SecureStreamSocket instance. This however cannot
work due to intensive using of copy-constructors without preserving the
original class type. Instead, the underlying SocketImpl pointer (which is
still polymorhic) is copied. Thus, the fixed implementation just ask for
the SocketImpl * and converts it to the SecureStreamSocketImpl *. The result
is much simplier :).

Fixes: abb76bc26dee ("SocketGatewayPeerVerifierFactory: new interface class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>