RelaxedGatewayPeerVerifier: refactor to be only a wrapper
The RelaxedGatewayPeerVerifier and its factory are now just wrappers.
They wrap any other GatewayPeerVerifier and SocketGatewayPeerVerifierFactory
and relax their verification operations to be logging-only and accepting all
peers.

This way, it is possible to test any GatewayPeerVerifier without
accidently denying access for gateways. The relaxed implementation is
considered for testing purposes only.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>