db: update to use splitted queries to access sensor_history
The strategy of fetching data from database is now decided on the
application level. The behaviour was left unchanged, if the interval
is below 5 seconds, no aggregation is performed and the system
returns raw data. When the interval is greater then 5 seconds,
we perform aggregation and return average values, max values or min
values based on the setting.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>