SqitchInitializer: fix handling paths containing a dot
If a SqitchChange name contains a dot, the Poco::Path would be
confused and use the part after the dot as the extension of the
path. Thus, when calling Poco::Path::setExtension, we remove part
of the particular change name and point to an invalid file path.

E.g. beeeon.gateway/create with extension "sql" should be transformed
into beeeon.gateway/create.sql. Prior to this fix, the transformation
would emit beeeon.sql instead ("gateway/create" is treated as the
extension and overriden).

Fixes: 0f29862e2a9d ("SqitchInitializer: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>