JSONDeviceDeserializer: apply refresh time with stricter rules
When accepting a request to change the refresh time from clients,
we must be careful with the refresh time semantics. If a device
does not support refresh time than we must not change it. Same
is the opposite. If the refresh time is supported by a device,
we must not disable it.

We consider that refresh >= 0 is updatable while refresh < 0 is
non-updatable (disabled at all).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>