PocoSQLSensorHistoryDao: handle invalid values as NULL
Curently, we do not check whether the sensor values (ValueAt, ModuleValue)
are valid or not. When inserting into database, the invalid values should
be inserted as NULL (and not NaN).

Also when retrieving them back, we should count on having NULLs there
and convert them to invalid values. On retrieval, the Poco "into" does
not work as expected with Nullable<double> (causes SIGSEGV). Thus,
change the retrieving of nullable values into the explicit approach
where we are accessing rows directly.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>