HTTPCustomHeadersFilter: new class
The HTTPCustomHeadersFilter is intended to serve HTTP headers not
provided by the application code. Headers that can be set to fixed
values are easily injectable in this way. The class is useful to
implement headers like:

* Strict-Transport-Security
* Content-Security-Policy
* X-Frame-Options
* X-XSS-Protection
* X-Content-Type-Options
* Referrer-Policy

and possibly others...

The headers are injected based on the request URI. The URI must match
the given URI pattern. Currently, the glob algorithm is used for
this purpose.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>