TypeInfoProvider: register types with no record in types.xml
TypeInfoProvider now registers also types which are not defined in
the types.xml. This allows to automatically include all types defined
in ModuleType::Type. The file types.xml is thus now used only for
extending the type definitions or overriding their properties that
are hard-coded.

Note that if a type is specified in the XML, it must be specified
with *all* its properties because the current implementation takes
either the XML definition or the hard-coded definition. Mixing of
both is not possible for now.

The types.xml is updated to avoid unnecessary definitions.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>