SensorHistoryRules: new class
SensorHistoryRules provides a way how to restrict history queries
(currently) based on the parameters: time range and time interval.

The time range specifies how long is the history period we are
interested in. Thus, it directly affects the data amount to work
with.

The time interval specifies the aggregation granularity. The bigger
the time interval is, the better aggregation is performed and less
data amounts are processed.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>