SensorXmlHandler: handle error conditions for getlogs
The getlogs is implemented as a streamed operation. Thus, if
an exception is thrown during data consuming, we cannot stop
the output stream as we have already sent some data.

However, if an exception is thrown from inside the service
and no data have been consumed yet, we can solve this situation.
The ValueConsumer implements begin(), single(), end() interface.
Until begin() is called, we can still report error. After the
begin() is called, exceptions must be swallowed and the output
finished gracefully.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>