RestValueConsumer: extract static method format()
For serializing purposes, extract just the formatting method
of ValueAt. This allows to share the serializing code of values
while breaking dependency of other serializing pruposes on the
implementation RestValueConsumer's ValueConsumer API.

When serializing just a single value, we do not want to call
begin(), end() as these calls can introduce unwanted output.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>