conf: fix sensor history rules to match legacy app needs
Raw access is requested by giving interval 0 from the legacy app.

The lowest range 0..1 s has only a single allowed interval "max"
which is unlikely to be asked (so it mostly leads to an error)
and if it is, we does not care as up to 1 s of data would most
likely lead to no result. However, this is a hack...

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Fixes: 4a389d2fa5e1 ("conf: fix sensor history rules to match legacy app needs")