OAuth2AuthProvider: rework to be generic and non-abstract
The OAuth2AuthProvider is now implemented via OAuth2CodeExchanger
and OAuth2UserFetcher components and thus it just converts the given
credentials to the appropriate form. The OAuth2AuthProvider is not
an abstract class anymore.

This change required to drop the existing OAuth2 implementations
GoogleAuthProvider and FacebookAuthProvider as they depend on it.
To replace them, create an instance of OAuth2AuthProvider and
use either GoogleOAuth2Provider or FacebookOAuth2Provider that
implement the OAuth2CodeExchanger and OAuth2UserFetcher interfaces.

The AuthRestHandler has been slighly changed to not advertise
OAuth2AuthProvider instance without clientId. Such situation can
occur if an OAuth2AuthProvider instance is not configured for
authorization code flow.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>