config: utilize PostgreStartup for testing
The PostgreStartup starts a new PostgreSQL instance with a clean
cluster and a default database named "postgres". Currently, there
is no login (this should be fixed) and the instance is listening
on 127.0.0.1 only.

The PostgreStartup is initialized and started by the testing-setup
LoopRunner that is always initialized early. However, it starts
(via autoStart) only when we are in the testing mode.

Also, the factory-testing.xml file is not installed for distribution
(same as for the testing-startup.ini).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>