PostgreStartup: call deletedb() from destructor instead of hook
The deletedb() should not be called from the cleanup hook as in that
case, it might delete the cluster data before the PostgreSQL is stopped.
This happens especially in the current setup when the PostgreSQL is
started from an early-initialized LoopRunner.

Thus, calling of deletedb() is moved into destructor where it works
as expected and in the right order.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>