PostgreStartup: change PGID of the postgres process
If the process 'postgres' has the same PGID as its parent (beeeon-server),
it is terminated when halting the server from shell (via ^C). This is
however an unwanted behaviour because the database server is halted too
early (before the stop() method is called).

By changing the PGID of the postgres process, it is not terminated by
the shell (that sends signals to the process group) but by the stop()
method and the server is terminated in the right order without any
reasons to cause failures (certain services might access database
during the termination).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>