PocoSQLSensorHistoryDao: remove failing values in insertMany
When calling insertMany() and some of the insertion fails, there was
now way how the upper layer could distinguish which values have failed
and which have succeeded.

The moduleValues are passed as non-const reference and it is thus possible
to remove failing values. Now, the caller might reiterate the vector and
it knows exactly the values that have succeeded.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>