GWMessageHandlerImpl: fix onReceived event generation
The event onReceived was generated even in the case when the actual
data insertion has failed. E.g. we can receive a malformed list of
values:

{
 ...
 "module_id": 0, "value": 30
 "module_id": 0, "value": 60
}

here with duplicated module 0. Such insertion would fail for the second
attempt to insert value for module 0 with the same timestamp. For such
data, we should generate the following event:

{
 ...
 "data": {"0": 30}
}

but we would have generated:

{
 ...
 "data": {"0": 30, "0": 60}
}

which is semantically wrong anyway. The event listener would see wrong
event and would either drop it all (loosing data) or process it in an
unexpected way.

By moving the event generation *after* calling to insertMany(), we
generate an event for only the values that have been successfully
inserted.

Fixes: e195d82e16c5 ("GWMessageHandlerImpl: fire events for SensorDataListeners")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>