DeviceUnpairHandler: handle both generic and specific result
The DeviceUnpairHandler now unpairs devices reported by the gateway
if it receives GatewayRPCUnpairResult instance. Otherwise, the legacy
way of handling responses to GatewayRPCResult is preserved.

If the device that was originally intended to be unpaired is not unpaired
but other devices are then we should report an event that the intended
device's unpair process has failed.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>