Sanitize: strict() can validate also a list of entries
The method strict() accepts parameter separator that can provide characters
to be used as separators. The input data can thus look e.g. like:

  - abcd0123
  - a,b,c,d
  - abcd,efgh,0123

The comma "," is normally not permitted for strict input. But with the
specified parameter separator, the input string is splitted first and
then its parts are sanitized.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>