SensorRestHandler: aggregation field is validated as a list of entries
The URI parameter aggregation can contain commas (,) as separators. This was
however not allowed by the Sanitize::strict() validator before. Now, the
validation accepts also the comma and thus it allows inputs like

 - avg
 - avg,min
 - min,max
 - avg,min,max

without any issues.

Tests are extended to make sure that it works and the response contains all
the requested data.

Fixes: 7f4b15e07945 ("SensorRestHandler: accept list of aggregators")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>