SensorHistoryServiceImpl: evaluate unknown values
When sensor data is received from a gateway for device that
has a module marked with the <unknown /> tag, its sensor value
is assumed to be in a raw format and it is converted via the
UnknownEvaluator.

Currently, it is not possible to support the same device type
with and without unknowns. If the <unknown /> tag is defined,
we *always* convert. This might be improved later by storing
that information into database.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>