ControlServiceImpl: evaluate values to raw when needed
When a user is trying to set a value of a controllable module
and the module is defined with the <unknown /> tag, the value
is converted to the raw form when sending to the gateway via
UnknownEvaluator.

The conversion is done for every controllable module with the
<unknown /> tag defined. This can be improved in the future by
recording the fact whether to use the unknown-conversion or not
into the database.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>