SocketReadyBlocker: optionally test a payload ping
The SocketReadyBlocker was previously only pinging a selected TCP
port. With this change, it is possible to configure an initial
request-response to be tested against the remote port. This makes
possible to test e.g. whether the remote server really provides the
expected service.

It can be configured to only send some request and have no expectations
about any response. Also vice-versa, to not send any request and expect
some response. And both sending a request and expecting some response.
Response can be checked as:

* whether its header matches
* whether whole response matches
* whether the response has the expected size

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>