PocoDefaultScriptExecutor: extract class from PocoDBInitializer
The PocoDBInitializer is now configurable and allows to set a different
PocoScriptExecutor implementation if needed.

The PocoDefaultScriptExecutor executes the given script as a single
statement, i.e. in the same way as the PocoDBInitializer did before.
It is set as a default implemention to stay backwards compatible.

Executor for PocoSimpleDBInitializer and PocoSqitchInitializer can
be configured via DI.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>