config: avoid INI keys to be changed for testing sqitch initialization
The initialization of the primary database by PocoSqitchInitializer is done
for testing purposes only. Testing is always done in the environment where
the deploy scripts are in the subproject db. There is no need to change this
configuration and the fact that the sqitch settings were present in the
testing-startup.ini only just proves this idea.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>