devices.xml: redefine Jablotron AC-88 outlet
The reworked Jablotron support on gateway changes behaviour of the AC-88
gadget. It is no longer a controllable device, it is just a sensor. This
new gadget is detected by its name containing "AC-88 (sensor)".

If a gateway is upgraded to new firmware, AC-88 devices has to be
re-paired on the server side to obtain the new type ID.

Device ID would be unchanged.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>