devices.xml: redefine Jablotron RC-88K remote control
The rework Jablotron support on gateway handles the RC gadget differently.
Both sides (two addresses) are now a single device. This version is thus
entirely incompatible as there is 1 controllable module more. The gadget
identifies itself by reporting "RC-86K (dual)".

If a gateway is upgraded to new firmware, this device has to be
re-paired on the server side to obtain the new type ID.

As the both sides of the RC-86K are now a single device, the device ID
of the right half would not work anymore. The device ID of the left
side would be the same, however, incompatible with the legacy definitions.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>