devices.xml: new device type PG(X/Y) Control
The reworked Jablotron support on gateway introduces a new type of
device that is connected with the PGX and PGY outputs of the Jablotron
system. The PGX and PGY are separate devices with the same definition.
Every pair process might associate a new Jablotron device with the
paired device (either PGX or PGY). Unpair does nothing at the gateway
level.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>