AsyncGatewayRPC: transform finalResultTimeout to defaultTimeout
The finalResultTimeout is logically wrong. If using it, the AsyncGatewayRPC
expects that any request to a gateway would last maximal finalResultTimeout.
However, many tasks have configurable timeout (listen, set-value) that is
usually longer.

Change finalResultTimeout to defaultTimeout. The defaultTimeout is used when
a request has no timeout defined. Otherwise, the request timeout is always
preferred.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>