t: gws: avoid explicit timeout specification for zmq
The ZMQConnection is an asynchronous interface. Thus, its operations
require to be protected by timeout.

It is now discouraged to explicitly override the timeout in particular
calls to pop_data() and leave it by using the default that can be
controlled globally. In specific situations, an explicit timeout can
be applicable.

The default timeout is updated to 10 to preserve the expected behaviour.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>